backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==4.2.16
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
sqlparse==0.5.1
tomli==2.0.2
typing-extensions==4.12.2
